Added proper array access to LuGI.  Now, when you access an array in a field or something, it will actually push that array object.  You can index it using integers or strings convertible to integers, as expected.  Beginning offset of those arrays is zero.
If you assign a table to an array field, it'll be converted to an array, and hopefully to one of the correct type.  It may be worthwhile to add an array constructor as well, just so you can create arrays in Lua, but that's at the discretion of people using LuGI.

Arrays support concatenation, slicing (via Array:Slice), sorting, getting length, etc.  There is also an additional method, :Type(), which'll return the type of the array.  May be handy, may not be, time will tell.

Made a few style changes about the code as well.